refactor(service): move provider tables down from cli, fix agent→cli inversion
